requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==39.0.0
Django==4.1.3
django-cors-headers==3.13.0
django-csp==3.7
django-redis==5.2.0
djangogrpcframework==0.2.1
djangorestframework==3.14.0
ez-setup==0.9
gunicorn==20.1.0
idna==3.4
jsonfield==3.1.0
kombu==5.2.4
mysqlclient==2.1.1
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
python-dotenv==0.21.0
pytz==2022.6
redis==4.4.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
stripe==4.2.0
tzdata==2022.7
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
